Código funcional, falta apenas documentação
