+++ src/sim.py
@@ -1,256 +1,173 @@
-"""
-Módulo com ferramentas para simulação e classes simuláveis.
-
-Classes
--------
-Particula
-    Classe base para todos os objetos da simulação em 2D.
-
-Variáveis
----------
-G = 1
-    Constante dagravitação universal. Sinta-se livre para alterá-la.
-
-Funções
--------
-limpar_objetos
-    Limpa a lista interna de objetos criados, deletando todos os objetos da simulação.
-iterar
-    Iteração simples usando método de Euler
-simular
-    Executa diversas iterações e rotorna uma lista de passos.
-
-"""
-
+from src import coisas as csa
 import numpy as np
+from matplotlib import pyplot as plt
+from matplotlib.animation import FuncAnimation, writers
 
-_objs = []  # lista global de objetos de todas as classes
-G = 1  # 6.6708e-11  # constante da gravitação universal
+formatos_suportados = ('3g2', '3pg', 'amv', 'asf', 'avi', 'dirac', 'drc', 'flv', 'gif', 'm4v', 'mp2', 'mp3', 'mp4',
+                       'mjpeg', 'mpeg', 'mpegets', 'mov', 'mkv', 'mxf', 'mxf_d10', 'mxf_opatom', 'nsv', 'null', 'ogg',
+                       'ogv', 'rm', 'roq', 'vob', 'webm')
 
 
-class Particula:
-    """
-    Classe base para todos os objetos da simulação em 2D.
-    Contem apenas posição inicial, velocidade inicial e massa de modo que não há colisões.
-    Um ponteiro no final da lista `_objs` é adicionado em toda instânciação, apontando para ele.
+class Sim:
+    def __init__(self, herdar=None):
+        if isinstance(herdar, Sim):  # possibilita herdar as configurações de outra sdimulação
+            self.objs = herdar.objs
+            self.dados = herdar.dados
+            self.tempos = herdar.tempos
+            self.configs = herdar.configs
+            self.h = herdar.h
+        else:
+            self.objs = []  # lista de objetos inclusos na simulação
+            self.dados = []  # dados gerados
+            self.tempos = []  # insatantes de cada passo
+            self.h = 0.01  # passo da simulação (padrão como 0.01)
+            self.configs = {'estilo': 'dark_background',  # estilo de fundo
+                            'seguir': -1,  # objeto seguido (negativo para nenhum)
+                            # todo: criar forma de seguir ponteiro, além de índice
+                            'xlim': (-5, 5), 'ylim': (-5, 5),
+                            'fps': 30, 'vel': 1,  # limites de enquadramento
+                            'G': 1}  # Constante da gravitação universal; real = 6.6708e-11; 0 para sem gravidade
+            # dicionário de configurações da simulação
 
-    Atributos
-    ---------
-    s : ndarray
-        Vetor posição da partícula.
-    v : ndarray
-        Vetor velocidade da partícula.
-    m : int ou float
-        Massa da partícula.
-    index : int
-        Indíce do objeto na lista `_objs`.
+    def add_obj(self, *args):
+        classes = [csa.Particula] + csa.Particula.__subclasses__()  # litsa de todas as classes suportadas na simulação
+        add = []  # lista que reune todos os objetos que serão adicionados
+        for a in args:
+            try:  # testa se o argumento é iterável ou um objeto
+                iter(a)
+                for obj in a:  # adiciona todos os objetos do iterável à lista `add`
+                    add.append(obj)
+            except TypeError:
+                add.append(a)  # adiciona o objeto à lista `add`
 
-    Métodos
-    -------
-    ar()
-        Retorna vetor aceleração resutante.
-    em_orbita(s, m=1.0, sentido=True)
-        Retorna uma partícula em órbita na posição e massa especificada.
+        for obj in add:  # testa e adiciona todos os objetos reunidos
+            for c in classes:
+                if isinstance(obj, c):  # checa se o objeto é de uma classe suportada
+                    self.objs.append(obj)
+                    obj.sim = self
+                    obj._def_sats()
+                else:
+                    print(f'{obj}, não é um objeto simulável')
 
-    """
-    def __init__(self, s=[0, 0], v=[0, 0], m=1.0):
-        """
-        Parâmetros
-        ----------
-        s : iterável (lista, tupla, ndarray, etc.) de formato (2,), padrão=[0,0]
-            Posição da partícula no início da simulação.
-            Independente do tipo de iterável, o código transforma em ndarray.
-        v : iterável (lista, tupla, ndarray, etc.) de formato (2,), padrão=[0,0]
-            Velocidade da partícula no início da simulação.
-            Independente do tipo de iterável, o código transforma em ndarray.
-        m : int ou float, padrão=1.0
-            Massa da partícula.
-        """
-        self.s = np.array(s)  # vetor de posição
-        self.v = np.array(v)  # vetor de velocidade
-        self.m = m  # massa do objeto
-        self.index = len(_objs)  # armazena o endereço do próprio objeto na lista _objs
-        _objs.append(self)  # adiciona o objeto à lista de pointers
-
-        # todo: adicionar opção de customizar cores
-
-    def ar(self):  # método para retornar o vetor de acleração resultante
-        """
-        Calcula o vetor aceleração resultante. No momento apeans a aceleração gravitacional é calculada
-        comparando, a partir da distância entre o objeto atual e todos os outros.
-
-        Parâmetros
-        ----------
-        Nenhum
-
-        Retorna
-        -------
-        ar : ndarray
-            Vetor 2D com a aceleração resultante no objeto.
-
-        Notas
-        -----
-        A equação usada para determinar a aceleração gravitacional de cada objeto é
-
-        .. math:: g = GM/(x^2 + y^2) * [x,y]/|[x,y]|
-
-        """
+    def _ar(self, obj):
+        gc = self.configs['G']  # pega o valor da constante da gravitação universal
         ar = np.zeros(2)
-        for n in _objs:
-            if n != self:
-                d = self.s - n.s  # vetor distancia
-                ar += - np.array(G * n.m / (d[0] ** 2 + d[1] ** 2) * d / np.linalg.norm(d))
+        for n in self.objs:
+            if n is not obj:
+                d = obj.s - n.s  # vetor distancia
+                ar += - np.array(gc * n.m / (d[0] ** 2 + d[1] ** 2) * d / np.linalg.norm(d))
                 # aceleração gravitacional, multiplicada pelo versor da distância
+                # aqui se colocaria outras forças a serem adicionadas a `ar`
         return ar
 
-        # todo: criar outros modelos de forças e formas de seleciona-los.
-        #  note que funções também são objetos e podem ser armazenadas dentro de listas. Portanto, pode-se fazer com que
-        #  ar retorne um dicionário com o nome de forças e strings como definições. Pode-se criar uma variável de forças
-        #  ativas dentro da classe. Dentre outras opções a se verificar futuramente. O que for feito deve ser a nível
-        #  desta classe. Objetos mais complexos que partículas devem herdar desta classe. Reformatar o projeto com isso
-        #  em mente.
+    def iterar(self):
+        h = self.h
+        s_list = []  # lista com posição dos objetos em cada iteração
+        for o in self.objs:  # calcula os estados para cada objeto
+            # atualiza o valor das variáveis do objeto o:
+            o.s = o.s + h * o.v
+            o.v = o.v + h * self._ar(o)
+            s_list.append(o.s)
+        return np.array(s_list)
 
-    def em_orbita(self, s, m=1.0, sentido=True):  # cria automaticamente um novo objeto em velocidade orbital
-        """
-        Cria um novo objeto da classe Particula com velocidade orbital em relação à instânia que chamou este método.
+    def simular(self, t, h=0.01):
+        if len(self.objs) == 0:
+            raise ValueError('Nenhum objeto adicionado a simulação atual.')
+        t_hist = np.arange(0, t, h)  # cria uma lista de instantes no intervalo e passo definido
+        if self.tempos == []:  # se a lista tempos era vazia, sobscreve ela com t_hist
+            self.tempos = t_hist
+        else:  # senão, adiciona t_hist ao final de tempos
+            np.append(self.tempos, t_hist)
+        self.h = h  # atualiza o valor de passo utilizado
 
-        Parâmeteros
-        ----------
-        s : iterável de formato (2,)
-            Posição da partícula no início da simulação.
-            Independente do tipo de iterável, o código transforma em ndarray.
-        m : int o float, padrão=1.0
-            Massa da partícula.
-        sentido : bool, padrão=True
-            True definido como sentido horáiro e False como anti-horário.
+        print('Calculando {} iterações e {} interações.'.format(len(t_hist), len(t_hist) * len(self.objs) ** 2))
+        s_hist = []  # lista com as posições dos objetos durante toda a simulação
+        for _ in t_hist:  # loop de iterações em cada instante
+            s_hist.append(self.iterar())  # adiciona as posições do frame à lista de iterações
 
-        Retorna
-        -------
-        Particula
-            Retorna um objeto da classse partícula em velocidade orbital ao redor do objeto que chamou o método.
+        if self.dados == []:  # se a lista dados era vazia, sobrescreve ela com s_hist
+            self.dados = np.array(s_hist)
+        else:  # senão, adicona s_hist ao final de dados
+            dados = self.dados.tolist()  # transforma em lista para não perder o formato
+            dados.append(s_hist)
+            self.dados = np.array(dados)
 
-        """
-        # sentido=True -> horário, sentido=False -> anti-horário
-        s = np.array(s)
-        d = s - self.s  # vetor distância
-        mod_v = np.sqrt(G * self.m / abs(np.linalg.norm(d)))  # módulo da velocidade orbital
-        angle = np.arccos(d[0] / abs(np.linalg.norm(d)))
-        if sentido:
-            angle -= np.pi / 2  # calcula a inclinação do vetor v
-        else:
-            angle += np.pi / 2  # no senitdo horáiro ou anti-horáio
-        v = np.array([mod_v * np.cos(angle), mod_v * np.sin(angle)]) + self.v
-        # a adição da velocidade do objeto principal serve para anular a velocidade relativa
-        return Particula(s, v, m)
+    def animar(self, salvar_em=''):
+        # configurações
+        plt.style.use(self.configs['estilo'])
+        xlim = self.configs['xlim']
+        ylim = self.configs['ylim']
+        seguir = self.configs['seguir']
+        vel = self.configs['vel']
+        fps = self.configs['fps']
 
-
-def limpar_objetos():  # função para limpar a lista de objetos (reiniciar simulações)
-    """
-    Limpa a lista `_objs` apagando todos os objetos.
-
-    Notas
-    -----
-    Útil apenas para mais de uma simulação no mesmo código.
-
-    """
-    global _objs
-    for n in _objs:
-        del n
+        s_hist = self.dados
+        h = self.h
+        t_hist = self.tempos
+        t_total = t_hist[-1] + h  # retoma duração da simulação
+        dt = (1 / fps) * vel  # intervalo entre frames
+        print('Compilando vídeo. Duração: {}s, numero de frames: {}'.format(t_total / vel, int(t_total / dt)))
 
 
-def colisao():
-    pass  # função que terá as colisões entre objetos
+        def func_animar(f):  # gerador de função animar. f é o frame atual
+            t = f * dt  # instante atual
+            p = np.argmax(t_hist >= t)  # passo atual (primeiro instate após o frame atual)
+            pos = s_hist[p]  # posições no passo atual
 
+            plt.cla()  # limpa o plot anterior
+            plt.axis('scaled')
 
-def iterar(h=1/30):  # padrão definido para 30 FPS
-    """
-    Iteração básica de uma simulação usando método de Euler. Atualiza todos os valores dos objetos um a um.
+            if seguir < 0:  # configurações de limite responsivo
+                plt.xlim(xlim)  # limites fixos
+                plt.ylim(ylim)
+            elif seguir > len(self.objs):
+                raise IndexError('Objeto selecionado para seguir fora da lista de objetos. '
+                                 'Lembre-se que o índice começa em 0.')
+            else:
+                plt.xlim(xlim + pos[seguir, 0])  # limite atualizado de acordo com a posição do objeto
+                plt.ylim(ylim + pos[seguir, 1])
 
-    Parâmetros
-    ----------
-    h: float, padrão= 1/30
-        Tamanho do passo entre cada iteração em segundos.
+            plt.scatter(pos[:, 0], pos[:, 1])  # plota os pontos
 
-    Retorna
-    -------
-    ndarray
-        Lista de vetores com posições de cada objeto. A ordem dos vetores correpsonde ao `.index` do objeto.
+            # todo: ajustar função para configurar de plot por objeto (cores e raios próprios)
+            # todo: adicionar acompanhar centro de massa
+            # todo: criar limites adaptáveis
 
-    Notas
-    -----
-    Quanto menor o valor de `h` maior precisão/reslução da simulação.
-    Um `h` negativo implica em uma simulação voltando no tempo, teoricamente.
+        anim = FuncAnimation(plt.gcf(), func_animar,
+                             frames=int(t_total / dt), interval=1000 * dt)  # faz o loop de animação
+        if salvar_em != '':
+            formato = salvar_em.split('.')[-1]  # retorna o texto após o último ponto do diretório
+            salvar = True
+            if formato not in formatos_suportados:  # checa se o formato está na lista de suportados
+                r = input(f'Este formato de vídeo, \'.{formato}\' pode não ser suportado.'
+                          f' Tentar mesmo assim? [S/N]\n->')
+                if r != 'S':
+                    print('Ok, então o vídeo não será salvo.')
+                    salvar = False
 
-    """
-    s_list = []  # lista com posição dos objetos em cada iteração
-    for i in _objs:  # calcula os estados para cada objeto
-        # atualiza o valor das variáveis do objeto i:
-        i.s = i.s + h * i.v
-        i.v = i.v + h * i._ar()
-        s_list.append(i.s)
-        colisao()
-    return np.array(s_list)
+            try:
+                if salvar:  # não tenta salvar a animação se o usuário esitir, acima
+                    writer = writers['ffmpeg']
+                    escritor = writer(fps=fps)
+                    anim.save(salvar_em, escritor)  # salva a animação usando ffmpeg
+                    print(f'Vídeo salvo em {salvar_em}')
+            except UnicodeDecodeError:
+                raise UserWarning('Este formato de vídeo não é suportado.'
+                                  '\nTente acessar simul.formatos_suportados para ver uma série de opções.'
+                                  '\nSugerido: .mp4')
+                # isto serve apenas para substuir o erro do matplolib que é ilegível
 
+        print('Exibindo...')
+        plt.show()
+        plt.close()
 
-def simular(t, h=1/30):  # função com as iterações em loop do sistema. Padrão definido para 30 FPS
-    """
-    Executa diversas iterações e retorna um histórico delas.
-
-    Parâmetros
-    ----------
-    t : int ou float
-        Tempo total da simulação em segundos.
-    h: float, padrão= 1/30
-        Tamanho do passo entre cada iteração em segundos.
-
-    Retorna
-    -------
-    s_hist : ndarray
-        Vetor com cada elemento sendo vetores posições de cada objeto em cada iteração
-    t_hist : ndarray
-        Vetor com o instante de cada iteração.
-
-    Notas
-    -----
-    Essa função retorna uma matriz de três dimensões, sendo a primeira coordenada uma lista com cada iteração;
-    a segunda, uma lista de vetores posição de cada objeto por iteração; e a terceira a devida coordenada `x` ou `y`.
-    Assim, em `s_hist[i,o,d]` temos que `i` é o índice de cada Iteração;
-    `o`, o índice de cada Objeto na dada iteração; e `d` a Direção/coordenada de cada objeto em cada iteração.
-
-    Ver também
-    ----------
-    iterar()
-
-    """
-    t_hist = np.arange(0, t, h)  # cria uma lista de instantes no intervalo e passo definido
-    print('Calculando {} iterações e {} interações.' .format(len(t_hist), len(t_hist) * len(_objs) ** 2))
-    s_hist = []  # lista com as posições dos objetos durante toda a simulação
-    for tn in t_hist:  # loop de iterações em cada instante
-        s_hist.append(iterar(h))  # adiciona as posições do frame à lista de iterações
-    return np.array(s_hist), t_hist
-
-
-if __name__ == '__main__':
-    from src import ani
-
-    sis = input('Selecione sistema de 3 corpos de exemplo: \n1 - Infinito \n2 - Com órbita externa \n->')
-    if sis == '1':
-        # sistema de 3 corpos formato infinito; G = 1
-        a = 0.3471128135672417
-        b = 0.532726851767674
-        Particula(s=[-1, 0], v=[a, b])
-        Particula(s=[1, 0], v=[a, b])
-        Particula(s=[0, 0], v=[-2 * a, -2 * b])
-    elif sis == '2':
-        # um sistema de 3 corpos; G = 1
-        Particula(s=[-1, 0], v=[0, -0.3660350371], m=2)
-        Particula(s=[1.254953728, 0], v=[0, 0.4593570344], m=0.5)
-        Particula(s=[2.745046272, 0], v=[0, 1.004783114], m=0.5)
-    else:
-        raise ValueError('Poxa! Era 1 ou 2, não tinha muita margem para erro, o que você botou aqui?')
-
-    ani.animar_sim(simular(10, 0.01), xlim=(-2, 2), ylim=(-2, 2), vel=1)
-
-
-# todo: criar função de delta-v numa determinada direção, bem como função de teleporte de objeto
-# todo: criar nova classe Esfera, adicionado orientação, torque, raio e colisões aos objetos
+    def reset(self):  # reseta a simulação, apagando dados e objetos
+        self.dados = []
+        self.tempos = []
+        self.objs = []
+        self.configs = {'estilo': 'dark_background',  # estilo de fundo
+                        'seguir': -1,  # objeto seguido (negativo para nenhum)
+                        'xlim': (-5, 5), 'ylim': (-5, 5),
+                        'fps': 30, 'vel': 1,  # limites de enquadramento
+                        'G': 1}  # Constante da gravitação universal; real = 6.6708e-11; 0 para sem gravidade
+        self.h = 0.01
